+++ src/app/_services/providers.tsx
@@ -3,7 +3,7 @@
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
 import { useState } from 'react';
-import { AuthProvider } from '@/app/services/auth-provider';
+import { AuthProvider } from '@/app/_services/auth-provider';
 
 type ProvidersProps = {
   children: React.ReactNode;
